+++ common/src/config/GameConfig.cpp
@@ -208,14 +208,6 @@ bool GameConfig::visit_vars(T&& visitor, bool is_save)
     result &= visitor(dash_faction_key, "Multi Ricochet", multi_ricochet);
     result &= visitor(dash_faction_key, "Unlimited Semi Auto", unlimited_semi_auto);
     result &= visitor(dash_faction_key, "Use Gaussian Bullet Spread", gaussian_spread);
-    result &= visitor(dash_faction_key, "World HUD CTF", world_hud_ctf);
-    result &= visitor(dash_faction_key, "World HUD Overdraw", world_hud_overdraw);
-    result &= visitor(dash_faction_key, "World HUD Big Text", world_hud_big_text);
-    result &= visitor(dash_faction_key, "World HUD Damage Numbers", world_hud_damage_numbers);
-    result &= visitor(dash_faction_key, "World HUD Spectate Player Labels", world_hud_spectate_player_labels);
-    result &= visitor(dash_faction_key, "World HUD Team Player Labels", world_hud_team_player_labels);
-    result &= visitor(dash_faction_key, "Play Hit Sounds", play_hit_sounds);
-    result &= visitor(dash_faction_key, "Play Taunt Sounds", play_taunt_sounds);
 
     return result;
 }